requirements.txt
Flask==2.3.2
waitress==2.1.2
python-dotenv==1.0.0
gunicorn==20.1.0
SQLAlchemy==2.0.23
pyodbc==4.0.39